Move FilterGroup typedef file to parent directory
