+++ src/gen3-ui-component/components/filters/typedef.js
@@ -1,4 +1,4 @@
-/// <reference path="../../../../GuppyComponents/typedef.js" />
+/// <reference path="../../../GuppyComponents/typedef.js" />
 
 /** @typedef {{ [option: string]: boolean }} OptionFilterStatus */
